# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inexxio ECS API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/erp/articles": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "List Articles",
        "operationId": "list_articles_api_v1_erp_articles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArticleResponse"
                  },
                  "type": "array",
                  "title": "Response List Articles Api V1 Erp Articles Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Create Article",
        "operationId": "create_article_api_v1_erp_articles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/erp/articles/{object_id}": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "Get Article",
        "operationId": "get_article_api_v1_erp_articles__object_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Object Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "articles"
        ],
        "summary": "Update Article",
        "operationId": "update_article_api_v1_erp_articles__object_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Object Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/erp/articles/{object_id}/deactivation-impact": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "Deactivation Impact",
        "description": "Wirkungsanalyse vor Inaktiv/Ersetzen: mitbetroffene Artikel, laufende\nAufträge, Lagerbestand.",
        "operationId": "deactivation_impact_api_v1_erp_articles__object_id__deactivation_impact_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Object Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeactivationImpact"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/erp/articles/{object_id}/deactivate": {
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Deactivate Article Endpoint",
        "operationId": "deactivate_article_endpoint_api_v1_erp_articles__object_id__deactivate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Object Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeactivateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/erp/articles/{object_id}/replace": {
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Replace Article Endpoint",
        "description": "Ersetzen: Duplikat als Entwurf anlegen, verknüpfen, Original inaktiv setzen.\nLiefert den **neuen** Artikel zurück (zum Anpassen).",
        "operationId": "replace_article_endpoint_api_v1_erp_articles__object_id__replace_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Object Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeactivateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/erp/articles/{object_id}/instances": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "List Article Instances",
        "description": "Bestand des Artikels: alle Bestands-Instanzen (Reiter «Bestand»).",
        "operationId": "list_article_instances_api_v1_erp_articles__object_id__instances_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Object Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InstanceResponse"
                  },
                  "title": "Response List Article Instances Api V1 Erp Articles  Object Id  Instances Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArticleCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "serialization": {
            "type": "string",
            "title": "Serialization"
          },
          "size": {
            "type": "string",
            "title": "Size"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Weight Kg"
          },
          "material": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material"
          },
          "cad_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cad Url"
          },
          "surface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "supplier_article_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Article Number"
          },
          "min_order_qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Order Qty"
          },
          "safety_stock": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Stock"
          }
        },
        "type": "object",
        "required": [
          "name",
          "unit",
          "serialization",
          "size",
          "weight_kg"
        ],
        "title": "ArticleCreate",
        "description": "Anlage eines Artikels über den '+'-Button. Status startet immer als 'draft'."
      },
...
      "ArticleResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "object_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "serialization": {
            "type": "string",
            "title": "Serialization"
          },
          "size": {
            "type": "string",
            "title": "Size"
          },
          "weight_kg": {
            "type": "string",
            "title": "Weight Kg"
          },
          "material": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material"
          },
          "cad_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cad Url"
          },
          "surface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "supplier_article_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Article Number"
          },
          "min_order_qty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Order Qty"
          },
          "safety_stock": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Stock"
          },
          "landed_unit_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Landed Unit Cost"
          },
          "computed_weight_kg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Computed Weight Kg"
          },
          "replaced_by_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replaced By Id"
          },
          "replaces_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replaces Id"
          },
          "unit_cost_low": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Cost Low"
          },
          "unit_cost_high": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Cost High"
          },
          "lead_time_days_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Time Days Low"
          },
          "lead_time_days_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Time Days High"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "object_id",
          "status",
          "name",
          "unit",
          "serialization",
          "size",
          "weight_kg",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "ArticleResponse"
      },
      "ArticleUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "serialization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serialization"
          },
          "size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "material": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material"
          },
          "cad_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cad Url"
          },
          "surface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "supplier_article_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Article Number"
          },
          "min_order_qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Order Qty"
          },
          "safety_stock": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Stock"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "expected_updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Updated At"
          }
        },
        "type": "object",
        "title": "ArticleUpdate",
        "description": "Teil-Update aus dem Detailfenster. Alle Felder optional."
      },
...
      "DeactivateRequest": {
        "properties": {
          "orders_mode": {
            "type": "string",
            "title": "Orders Mode",
            "default": "phase_out"
          }
        },
        "type": "object",
        "title": "DeactivateRequest",
        "description": "Optionen beim Inaktiv-Setzen/Ersetzen."
      },
      "DeactivationImpact": {
        "properties": {
          "articles": {
            "items": {
              "$ref": "#/components/schemas/ImpactArticle"
            },
            "type": "array",
            "title": "Articles",
            "default": []
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/ImpactOrder"
            },
            "type": "array",
            "title": "Orders",
            "default": []
          },
          "stock": {
            "type": "integer",
            "title": "Stock",
            "default": 0
          }
        },
        "type": "object",
        "title": "DeactivationImpact",
        "description": "Wirkungsanalyse (Dry-Run) vor dem Inaktiv-Setzen/Ersetzen eines Artikels."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}